+++ src/test/scala/de/upb/cs/uc4/hyperledger/tests/contracts/ExaminationRegulationErrorTests.scala
@@ -1,4 +1,4 @@
-package de.upb.cs.uc4.hyperledger.tests
+package de.upb.cs.uc4.hyperledger.tests.contracts
 
 import de.upb.cs.uc4.hyperledger.connections.traits.ConnectionExaminationRegulationTrait
 import de.upb.cs.uc4.hyperledger.testBase.TestBase
